// packages/components/src/api/Runs.ts
// Generated by https://quicktype.io

export interface Run {
    RunID:         string;
    Configuration: Configuration;
    CurrentState:  string;
}

export interface Configuration {
    ExpectedData:   ExpectedData;
    ExecutorConfig: ExecutorConfig;
    Processor:      Processor | null;
}

export interface ExecutorConfig {
    MaxWorkers:          number;
    SendDownstream:      boolean;
    DownstreamLocations: DownstreamLocation[];
}

export interface DownstreamLocation {
    Hostname: string;
    ProcID:   string;
    RunID:    string;
}

export interface ExpectedData {
    NumChunks: number;
}

export interface Processor {
    test: string;
}

export type Runs = Run[]